Moved drag related elements into their own namespace.
