+++ AnnoMapEditor/UI/Controls/Dragging/DragEndedEvent.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AnnoMapEditor.UI.Controls
+namespace AnnoMapEditor.UI.Controls.Dragging
 {
     public delegate void DragEndedEventHandler(object? sender, DragEndedEventArgs e);
