+++ AnnoMapEditor/UI/Controls/Dragging/DraggingViewModel.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AnnoMapEditor.UI.Controls
+namespace AnnoMapEditor.UI.Controls.Dragging
 {
     public abstract class DraggingViewModel : ObservableBase
     {
