+++ AnnoMapEditor/UI/Controls/MapTemplates/PlayableAreaControl.xaml
@@ -1,20 +1,20 @@
-﻿<controls:DraggingControl x:Class="AnnoMapEditor.UI.Controls.MapTemplates.PlayableAreaControl"
+﻿<dragging:DraggingControl x:Class="AnnoMapEditor.UI.Controls.MapTemplates.PlayableAreaControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:AnnoMapEditor.UI.Controls.MapTemplates"
-             xmlns:controls="clr-namespace:AnnoMapEditor.UI.Controls"
              xmlns:converters="clr-namespace:AnnoMapEditor.UI.Converters"
+             xmlns:dragging="clr-namespace:AnnoMapEditor.UI.Controls.Dragging"
              mc:Ignorable="d" 
              d:DataContext="{d:DesignInstance Type=local:PlayableAreaViewModel}"
              d:DesignHeight="400" d:DesignWidth="400">
 
-    <controls:DraggingControl.Resources>
+    <dragging:DraggingControl.Resources>
         <converters:BoolToVisibility x:Key="VisibleOnTrue" />
         <converters:BoolToVisibility x:Key="VisibleOnFalse" OnTrue="Collapsed" OnFalse="Visible" />
         <converters:IntToGridLengthConverter x:Key="IntToGridLength" />
-    </controls:DraggingControl.Resources>
+    </dragging:DraggingControl.Resources>
     
     <Grid>
         <Grid.ColumnDefinitions>
@@ -113,4 +113,4 @@
             </Border>
         </Grid>
     </Grid>
-</controls:DraggingControl>
+</dragging:DraggingControl>
