+++ AnnoMapEditor/UI/Controls/MapTemplates/PlayableAreaViewModel.cs
@@ -1,10 +1,7 @@
 ﻿using AnnoMapEditor.MapTemplates.Models;
+using AnnoMapEditor.UI.Controls.Dragging;
 using AnnoMapEditor.Utilities;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AnnoMapEditor.UI.Controls.MapTemplates
 {
